[SQUASHED] Add in roomservice
thx to cm, omni and pa

commit history:
=================

[*]roomservice: Auto-remove duplicated projects

roomservice has improved in the sense that once a duplicate path is
found, a <remove-project> tag will be added into roomservice.xml

Change-Id: If3e1dc672d25996a0cd19628ece24c010e213adf

[*]roomservice: Include forked repositories in device search

The github api defaults to not including forked repos in search results.
This causes roomservice to fail to find a device repo if it is a fork (I.E. maguro).
This commit adds the necessary argument to include forked repositories in the results
and fixes maguro syncing.

Change-Id: I50ec96316c3208e977d659a1dd90205a2b3d85b3
Signed-off-by: Evan Anderson <evan1124@gmail.com>

[*]roomservice: Fix device syncing

Change-Id: Id4a59d74531a4ea1744f845fbe65b43debf950dd
Signed-off-by: Evan Anderson <evan1124@gmail.com>

[*]Switch to OmniRom's roomservice

All credit goes to Omni for this.

PS2: Update defaults to use with PA
PS3: Give proper authorship
PS4: Update dependency file name
PS5: Set default_team_rem to github
PS6: Revert PS5. Depends on manifest change

Change-Id: I77fe0e958a3da5a3d65d0bf8c929fed2b13115bd
Signed-off-by: Evan Anderson <evan1124@gmail.com>

[*]roomservice: fix missing revision line

Change-Id: I727f3a6ab1345b32b61b1c1bda31f03008082c96
Signed-off-by: Evan Anderson <evan1124@gmail.com>

[*]roomservice: Fix adding to manifest

Python is indent-sensitive, so needed to indent some lines.

Change-Id: Iaa360754038e4966fb76eb264e7c3195209735a1
Signed-off-by: Evan Anderson <evan1124@gmail.com>

Contributers:
@Ryuinferno
@Cybojenix

Change-Id: I91ee64727edf80cbe2cce7934de0adcf813b55a6

envsetup: Correct roomservice logic

Calling roomservice properly relied on the exit status of check_device.
Since the aosp device check is in between, we need to save the exit status
to a local variable for the logic to work properly.

Change-Id: I5f294565259f5e88a953602500a9cfbe9bdc590b
Signed-off-by: Evan Anderson <evananderson@aospa.co>

build: Fix roomservice